+++ mapblockaccessor/update.go
@@ -8,6 +8,9 @@ import (
 )
 
 func (a *MapBlockAccessor) Update(pos *types.MapBlockCoords, mb *mapparser.MapBlock) {
+	if a.maxcount <= 0 {
+		return
+	}
 	key := getKey(pos)
 	cacheBlockCount.Inc()
 	a.blockcache.Set(key, mb, cache.DefaultExpiration)